Login com Github e Lazy/Suspense em IA e Calculator
